Moved and renamed 2D and 3D jacobians for isoparametric interpolations.
