+++ ISAAR.MSolve.FEM/Elements/ContinuumElement3D.cs
@@ -8,6 +8,7 @@ using ISAAR.MSolve.FEM.Entities;
 using ISAAR.MSolve.FEM.Interfaces;
 using ISAAR.MSolve.FEM.Interpolation;
 using ISAAR.MSolve.FEM.Interpolation.GaussPointExtrapolation;
+using ISAAR.MSolve.FEM.Interpolation.Jacobians;
 using ISAAR.MSolve.FEM.Materials;
 using ISAAR.MSolve.Materials;
 using ISAAR.MSolve.Numerical.LinearAlgebra;
@@ -70,7 +71,7 @@ namespace ISAAR.MSolve.FEM.Elements
             {
                 Matrix2D shapeFunctionMatrix = BuildShapeFunctionMatrix(shapeFunctions[gaussPoint]);
                 Matrix2D partial = shapeFunctionMatrix.Transpose() * shapeFunctionMatrix;
-                var jacobian = new Jacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
+                var jacobian = new IsoparametricJacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
                 double dA = jacobian.Determinant * gaussPoint.Weight;
                 mass.AxpyIntoThis(partial,dA);
             }
@@ -88,7 +89,7 @@ namespace ISAAR.MSolve.FEM.Elements
             double area = 0;
             foreach (var gaussPoint in QuadratureForConsistentMass.IntegrationPoints)
             {
-                var jacobian = new Jacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
+                var jacobian = new IsoparametricJacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
                 area += jacobian.Determinant * gaussPoint.Weight;
             }
 
@@ -109,7 +110,7 @@ namespace ISAAR.MSolve.FEM.Elements
             foreach (var gaussPoint in QuadratureForStiffness.IntegrationPoints)
             {
                 Matrix2D constitutive = (Matrix2D) materialsAtGaussPoints[gaussPoint].ConstitutiveMatrix;
-                var jacobian = new Jacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
+                var jacobian = new IsoparametricJacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
                 Matrix2D shapeGradientsCartesian =
                     jacobian.TransformNaturalDerivativesToCartesian(shapeGradientsNatural[gaussPoint]);
                 Matrix2D deformation = BuildDeformationMatrix(shapeGradientsCartesian);
@@ -230,7 +231,7 @@ namespace ISAAR.MSolve.FEM.Elements
             {
                 GaussPoint3D gaussPoint = QuadratureForStiffness.IntegrationPoints[i];
                 IMatrix2D constitutive = materialsAtGaussPoints[gaussPoint].ConstitutiveMatrix;
-                var jacobian = new Jacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
+                var jacobian = new IsoparametricJacobian3D(Nodes, shapeGradientsNatural[gaussPoint]);
                 Matrix2D shapeGrandientsCartesian =
                     jacobian.TransformNaturalDerivativesToCartesian(shapeGradientsNatural[gaussPoint]);
                 Matrix2D deformation = BuildDeformationMatrix(shapeGrandientsCartesian);
